v(2.0.2.19) update PowerPostLib.hlsl,half to float(half is not enough)

+++ PowerPostPrj/PowerPost/Resources/Shaders/PowerPostLib.hlsl
@@ -7,22 +7,22 @@
 #include "Packages/com.unity.render-pipelines.universal/ShaderLibrary/DeclareDepthTexture.hlsl"
 #include "../../../../../PowerShaderLib/Lib/CoordinateSystem.hlsl"
 
-half4 SampleBox(TEXTURE2D_PARAM(tex,state), half4 texel, half2 uv, half delta) {
-	half2 p = texel.xy * delta;
-	half4 c = SAMPLE_TEXTURE2D(tex, state, uv + half2(-1, -1) * p);
-	c += SAMPLE_TEXTURE2D(tex, state, uv + half2(1, -1) * p);
-	c += SAMPLE_TEXTURE2D(tex, state, uv + half2(-1, 1) * p);
-	c += SAMPLE_TEXTURE2D(tex, state, uv + half2(1, 1) * p);
+float4 SampleBox(TEXTURE2D_PARAM(tex,state), float4 texel, float2 uv, float delta) {
+	float2 p = texel.xy * delta;
+	float4 c = SAMPLE_TEXTURE2D(tex, state, uv + float2(-1, -1) * p);
+	c += SAMPLE_TEXTURE2D(tex, state, uv + float2(1, -1) * p);
+	c += SAMPLE_TEXTURE2D(tex, state, uv + float2(-1, 1) * p);
+	c += SAMPLE_TEXTURE2D(tex, state, uv + float2(1, 1) * p);
 
 	return c * 0.25;
 }
 
-half4 SampleBox(TEXTURE2D_PARAM(tex,state),half4 texel,half2 uv, half delta,half sideWeight,half centerWeight) {
-    half2 p = texel.xy * delta;
-    half4 c = SAMPLE_TEXTURE2D(tex,state,uv + half2(-1,-1) * p) * sideWeight;
-    c += SAMPLE_TEXTURE2D(tex,state,uv + half2(1,-1) * p) * sideWeight;
-    c += SAMPLE_TEXTURE2D(tex,state,uv + half2(-1,1) * p) * sideWeight;
-    c += SAMPLE_TEXTURE2D(tex,state,uv + half2(1,1) * p) * sideWeight;
+float4 SampleBox(TEXTURE2D_PARAM(tex,state),float4 texel,float2 uv, float delta,float sideWeight,float centerWeight) {
+    float2 p = texel.xy * delta;
+    float4 c = SAMPLE_TEXTURE2D(tex,state,uv + float2(-1,-1) * p) * sideWeight;
+    c += SAMPLE_TEXTURE2D(tex,state,uv + float2(1,-1) * p) * sideWeight;
+    c += SAMPLE_TEXTURE2D(tex,state,uv + float2(-1,1) * p) * sideWeight;
+    c += SAMPLE_TEXTURE2D(tex,state,uv + float2(1,1) * p) * sideWeight;
     //c += SAMPLE_TEXTURE2D(tex,state,uv) * centerWeight;
     return c;
 }
@@ -31,10 +31,10 @@ half4 SampleBox(TEXTURE2D_PARAM(tex,state),half4 texel,half2 uv, half delta,half
 //----------------
 // Gaussian blur
 //----------------
-const static half gaussWeights[4]={0.00038771,0.01330373,0.11098164,0.22508352};
+const static float gaussWeights[4]={0.00038771,0.01330373,0.11098164,0.22508352};
 
-half4 GaussBlur(TEXTURE2D_PARAM(tex,sampler_tex),half2 uv,half2 offset,bool samplerCenter){
-    half4 c = 0;
+float4 GaussBlur(TEXTURE2D_PARAM(tex,sampler_tex),float2 uv,float2 offset,bool samplerCenter){
+    float4 c = 0;
     if(samplerCenter){
         c += SAMPLE_TEXTURE2D(tex,sampler_tex,uv) * gaussWeights[3];
     }
@@ -49,9 +49,9 @@ half4 GaussBlur(TEXTURE2D_PARAM(tex,sampler_tex),half2 uv,half2 offset,bool samp
     return c;
 }
 
-half4 SampleGaussian(TEXTURE2D_PARAM(tex,state), half2 texel, half2 uv) {
-	half4 c = GaussBlur(tex,state,uv,half2(texel.x,0),true);
-	c += GaussBlur(tex,state,uv,half2(0,texel.y),true);
+float4 SampleGaussian(TEXTURE2D_PARAM(tex,state), float2 texel, float2 uv) {
+	float4 c = GaussBlur(tex,state,uv,float2(texel.x,0),true);
+	c += GaussBlur(tex,state,uv,float2(0,texel.y),true);
 	return c;
 }
 
@@ -60,9 +60,9 @@ half4 SampleGaussian(TEXTURE2D_PARAM(tex,state), half2 texel, half2 uv) {
 // Dir blur
 //----------------
 #define DIR_BLUR_SAMPLES 6
-const static half dirBlurWeights[DIR_BLUR_SAMPLES] = {-0.03,-0.02,-0.01,0.01,0.02,0.03};
-half4 SampleDirBlur(TEXTURE2D_PARAM(tex,state),half2 uv,half2 dir){
-    half4 c = 0;
+const static float dirBlurWeights[DIR_BLUR_SAMPLES] = {-0.03,-0.02,-0.01,0.01,0.02,0.03};
+float4 SampleDirBlur(TEXTURE2D_PARAM(tex,state),float2 uv,float2 dir){
+    float4 c = 0;
     for(int i=0;i<DIR_BLUR_SAMPLES;i++){
         c += SAMPLE_TEXTURE2D(tex,state,(uv + dir * dirBlurWeights[i]));
     }
@@ -70,26 +70,26 @@ half4 SampleDirBlur(TEXTURE2D_PARAM(tex,state),half2 uv,half2 dir){
 }
 
 
-half Gray(half3 c){
-    return dot(half3(0.2,0.7,0.07),c);
+float Gray(float3 c){
+    return dot(float3(0.2,0.7,0.07),c);
 }
 
-half N21(half2 p){
-    return frac(sin(dot(p,half2(100,7890)))*12345);
+float N21(float2 p){
+    return frac(sin(dot(p,float2(100,7890)))*12345);
 }
 
-half N11(half c,half x=9.456){
+float N11(float c,float x=9.456){
     return frac(sin(c+x) * 12345.6789);
 }
 
 struct AttributesDefault{
-    half4 vertex:POSITION;
-    half2 texcoord:TEXCOORD;
+    float4 vertex:POSITION;
+    float2 texcoord:TEXCOORD;
 };
 
 struct VaryingsDefault{
-    half4 vertex:SV_POSITION;
-    half2 texcoord:TEXCOORD;
+    float4 vertex:SV_POSITION;
+    float2 texcoord:TEXCOORD;
 };
 
 /** fullscreen quad
@@ -99,7 +99,7 @@ struct VaryingsDefault{
 */
 VaryingsDefault VertDefault(AttributesDefault input){
     VaryingsDefault output;
-    output.vertex = half4(input.vertex.xy,0,1);
+    output.vertex = float4(input.vertex.xy,0,1);
     output.texcoord = input.texcoord;//(input.vertex.xy + 1) * 0.5;
     return output;
 }
@@ -111,12 +111,12 @@ VaryingsDefault VertDefault(AttributesDefault input){
     0
     -1 0 1 . 3
 */
-const static half2 TRIANGLE_VERTEIES[3] = {-1,-1,-1,3,3,-1};
-const static half2 TRIANGLE_UV[3] = {0,0,0,2,2,0};
+const static float2 TRIANGLE_VERTEIES[3] = {-1,-1,-1,3,3,-1};
+const static float2 TRIANGLE_UV[3] = {0,0,0,2,2,0};
 
 VaryingsDefault VertDefault1(uint vid:SV_VertexID){
     VaryingsDefault output;
-    output.vertex = half4(TRIANGLE_VERTEIES[vid],0,1);
+    output.vertex = float4(TRIANGLE_VERTEIES[vid],0,1);
     output.texcoord = TRIANGLE_UV[vid];
     return output;
 }
@@ -129,14 +129,14 @@ VaryingsDefault VertDefault1(uint vid:SV_VertexID){
     
 */
 #define OFFSET_COUNT 4
-static half2 offsets[OFFSET_COUNT] = {
-    half2(-1,0),
-    half2(0,1),
-    half2(1,0),
-    half2(0,-1)
+static float2 offsets[OFFSET_COUNT] = {
+    float2(-1,0),
+    float2(0,1),
+    float2(1,0),
+    float2(0,-1)
 };
 
-float CalcDepthDeltaAround(TEXTURE2D_PARAM(depthTex,state),float depth,float2 depthUV,half scale){
+float CalcDepthDeltaAround(TEXTURE2D_PARAM(depthTex,state),float depth,float2 depthUV,float scale){
     float delta = 0;
     for( int x = 0;x < OFFSET_COUNT ;x++){
         delta += depth - Linear01Depth(SAMPLE_TEXTURE2D(depthTex,state, depthUV + offsets[x] * scale).x,_ZBufferParams);
